[Bronze V] Title: 이상한 기호, Time: 108 ms, Memory: 14420 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,7 +7,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         int[] arr = Arrays.stream(br.readLine().split(" ")).mapToInt(Integer::parseInt).toArray();
-        int result = (arr[0] + arr[1]) * (arr[0] - arr[1]);
+        long result = (long) (arr[0] + arr[1]) * (arr[0] - arr[1]);
         bw.write(String.valueOf(result));
         bw.close();
     }
